Swipe Gesture recognizer for next question and LongTap for big image
The swipe gesture recognizer in QuestionViewController was added to
call the same method that the next question button calls.

The 4 answer buttons are hooked up to a touch down method that starts a
two second timer and displays the large image if the user does not
touchUpInside of the button. This can lead to problems if the user
touches down on the button but accidentally doesn’t touch up inside.